requirements.txt
aiogram
python-dotenv
requests
pymysql
pynput